requirements.txt
uvicorn~=0.29.0
fastapi~=0.110.0
SQLAlchemy~=2.0.29
pydantic~=2.6.4
pydantic_settings
pre-commit
psycopg2==2.9.9
asyncpg
redis~=5.0.3
alembic~=1.13.1
passlib~=1.7.4
bcrypt
PyJWT~=2.8.0
starlette~=0.36.3
loguru~=0.7.2
pytest~=8.1.1
pytest-asyncio
python-multipart
httpx~=0.27.0
uvicorn[standard]
